requirements.txt
anyio==3.6.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fastapi==0.85.2
fonttools==4.38.0
h11==0.14.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.0
starlette==0.20.4
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5